# resources/swagger/georchestra-signalement.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "Georchestra - Plugin Signalement API",
		"contact": {
			"email": "guillaume.morin@open-groupe.com"
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"host": "",
	"basePath": "/signalement",
...
	"schemes": [
		"https",
		"http"
	],
	"paths": {
...
		"/contextDescriptions": {
			"get": {
				"tags": [
					"contextDescriptions"
				],
				"summary": "Liste les desctiptions de contexte disponibles.",
				"description": "Retourne la liste des desctiptions de contexte existantes",
				"operationId": "searchContextDescriptions",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "query",
						"name": "description",
						"description": "Filtre de recherche sur la description",
						"required": false,
						"type": "string",
						"default": ""
					},
					{
						"in": "query",
						"name": "workflow",
						"description": "Filtre de recherche sur le workflow",
						"required": false,
						"type": "string",
						"default": ""
					},
					{
						"in": "query",
						"name": "offset",
						"description": "Index de début",
						"required": false,
						"type": "integer"
					},
					{
						"in": "query",
						"name": "limit",
						"description": "Nombre de résultats à retourner",
						"required": false,
						"type": "integer"
					},
					{
						"in": "query",
						"name": "sortExpression",
						"description": "Indications de tri",
						"required": false,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/ContextDescriptionPageResult"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			},
			"post": {
				"tags": [
					"contextDescriptions"
				],
				"summary": "Créer une desctiption de contexte.",
				"description": "Créer une desctiption de contexte telle que stipulée dans le body de la requête",
				"operationId": "createContextDescription",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "contextDescription",
						"description": "Description du contexte",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ContextDescription"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/ContextDescription"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			},
			"put": {
				"tags": [
					"contextDescriptions"
				],
				"summary": "Mise à jour d'une description de contexte.",
				"description": "Mise à jour d'une desctiption de contexte telle que stipulée dans le body de la requête, en fonction du nom de la description",
				"operationId": "updateContextDescription",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "contextDescription",
						"description": "Description du contexte",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ContextDescription"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/ContextDescription"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"404": {
						"description": "NOT FOUND"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/contextDescriptions/{name}": {
			"get": {
				"tags": [
					"contextDescriptions"
				],
				"summary": "Récupération d'une description de contexte.",
				"description": "Récupération d'une desctiption de contexte en fonction de son nom",
				"operationId": "getContextDescription",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "name",
						"description": "Nom de la description du contexte à récupérer",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/ContextDescription"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"404": {
						"description": "NOT FOUND"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			},
			"delete": {
				"tags": [
					"contextDescriptions"
				],
				"summary": "Suppression d'une description de contexte.",
				"description": "Suppression d'une desctiption de contexte en fonction de son nom",
				"operationId": "deleteContextDescription",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "name",
						"description": "Nom de la description du contexte à supprimer",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Bad input parameter"
					},
					"404": {
						"description": "NOT FOUND"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
...
		"/geographicareas": {
			"get": {
				"tags": [
					"geographicAreas"
				],
				"summary": "Récupération de la liste des zones géographiques disponibles",
				"description": "Récupération de la liste des zones géographiques disponibles, avec la possibilité de filtrer en fonction du nom",
				"operationId": "searchGeographicAreas",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "query",
						"name": "name",
						"description": "Nom ou partie du nom pour filtrer",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "offset",
						"description": "Index de début",
						"required": false,
						"type": "integer"
					},
					{
						"in": "query",
						"name": "limit",
						"description": "Nombre de résultats à retourner",
						"required": false,
						"type": "integer"
					},
					{
						"in": "query",
						"name": "sortExpression",
						"description": "Indications de tri",
						"required": false,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/GeographicAreaPageResult"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
		"/geographicareas/{id}": {
			"get": {
				"tags": [
					"geographicAreas"
				],
				"summary": "Récupération d'une zones géographique",
				"description": "Récupération d'une zones géographique en fonction de son id",
				"operationId": "getGeographicArea",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"description": "Id de la zone géographique",
						"required": true,
						"type": "integer",
						"format": "int64"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"$ref": "#/definitions/GeographicArea"
						}
					},
					"400": {
						"description": "Bad input parameter"
					},
					"404": {
						"description": "NOT FOUND"
					},
					"500": {
						"description": "Internal server error",
						"schema": {
							"$ref": "#/definitions/ApiError"
						}
					}
				}
			}
		},
...
	},
	"definitions": {
		"ApiError": {
			"type": "object",
			"properties": {
				"code": {
					"type": "string"
				},
				"label": {
					"type": "string"
				}
			}
		},
...
		"GeographicType": {
			"type": "string",
			"description": "Type géographic",
			"enum": [
				"POINT",
				"LINE",
				"POLYGON"
			]
		},
...
		"ContextType": {
			"type": "string",
			"description": "Type du context du signalement",
			"enum": [
				"THEMA",
				"LAYER"
			]
		},
...
		"ContextDescription": {
			"type": "object",
			"properties": {
				"contextType": {
					"type": "object",
					"$ref": "#/definitions/ContextType"
				},
				"geographicType": {
					"type": "object",
					"$ref": "#/definitions/GeographicType"
				},
				"name": {
					"type": "string"
				},
				"label": {
					"type": "string"
				},
				"processDefinitionKey": {
					"type": "string"
				},
				"revision": {
					"type": "integer"
				}
			}
		},
		"ContextDescriptionPageResult": {
			"type": "object",
			"properties": {
				"totalItems": {
					"type": "integer",
					"format": "int64"
				},
				"results": {
					"type": "array",
					"items": {
						"type": "object",
						"$ref": "#/definitions/ContextDescription"
					}
				}
			}
		},
...
		"GeographicArea": {
			"type": "object",
			"properties": {
				"id": {
					"type": "integer",
					"format": "int64"
				},
				"nom": {
					"type": "string"
				},
				"codeInsee": {
					"type": "string"
				}
			}
		},
		"GeographicAreaPageResult": {
			"type": "object",
			"properties": {
				"totalItems": {
					"type": "integer",
					"format": "int64"
				},
				"results": {
					"type": "array",
					"items": {
						"type": "object",
						"$ref": "#/definitions/GeographicArea"
					}
				}
			}
		},
...
	}
}